requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.0
Flask-Login==0.5.0
Flask-Bcrypt==1.0.1
gunicorn==20.1.0
flask 
flask-sqlalchemy 
mysql-connector-python